ai-course-code/chapter06_langchain_advanced/smart_content_processor/src/config.py
"""
配置管理
"""

import os
from pathlib import Path
from dotenv import load_dotenv

PROJECT_ROOT = Path(__file__).parent.parent
load_dotenv(PROJECT_ROOT / ".env")


class Config:
    """全局配置"""

    PROJECT_ROOT = PROJECT_ROOT

    # LM Studio 本地模型
    LM_STUDIO_BASE_URL = os.getenv("LM_STUDIO_BASE_URL", "http://localhost:1234/v1")
    LM_STUDIO_API_KEY = os.getenv("LM_STUDIO_API_KEY", "lm-studio")
    LM_STUDIO_MODEL = os.getenv("LM_STUDIO_MODEL", "qwen2.5-7b-instruct")

    # OpenAI（可选）
    OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
    OPENAI_API_BASE = os.getenv("OPENAI_API_BASE", "https://api.openai.com/v1")

    # 模型选择
    USE_LOCAL_MODEL = os.getenv("USE_LOCAL_MODEL", "true").lower() == "true"

    # 功能开关
    ENABLE_CACHE = True
    ENABLE_MONITOR = True

    # 性能告警阈值（秒）
    SLOW_REQUEST_THRESHOLD = 10


config = Config()
