ai-course-code/chapter06_langchain_advanced/smart_content_processor/requirements.txt
# 智能内容处理系统依赖
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
